Added the override to the launch script and node.
You can now override the calibration for GPS and baro with whatever
value is in the parameters. The parameters are update on every
calibration that does occur. This means that in flight you can
immeadiatley restart as long as the seed_estimator flag is passed to the
launch or run command.